+++ part4/tests/blog_api.test.js
@@ -1,14 +1,28 @@
 const mongoose = require("mongoose");
 const supertest = require("supertest");
+const bcrypt = require("bcrypt");
 const app = require("../app");
 const Blog = require("../models/blog");
+const User = require("../models/user");
 const { initialBlogs, nonExistingId, blogsInDb } = require("./api_test_helper");
-
 const api = supertest(app);
+
+let auth = {};
+let authUser = {};
+
+beforeAll(async () => {
+  await User.deleteMany({});
+  const passwordHash = await bcrypt.hash("sekret", 10);
+  const user = new User({ username: "root", passwordHash });
+  authUser = await user.save();
+
+  auth = await api.post("/api/login").send({ username: "root", password: "sekret" });
+  auth = auth.body;
+});
 
 beforeEach(async () => {
   await Blog.deleteMany({});
-  const blogModelList = initialBlogs.map((blog) => new Blog(blog));
+  const blogModelList = initialBlogs.map((blog) => new Blog({ ...blog, user: authUser._id }));
   const promiseArray = blogModelList.map((blogModel) => blogModel.save());
   await Promise.all(promiseArray);
 });
@@ -28,7 +42,7 @@ test("unique identifier property of blog is named id", async () => {
     likes: 12,
   };
 
-  const savedBlog = await api.post("/api/blogs").send(newBlog);
+  const savedBlog = await await api.post("/api/blogs").set("Authorization", `Bearer ${auth.token}`).send(newBlog);
   expect(savedBlog.body.id).toBeDefined();
 });
 
@@ -40,7 +54,7 @@ test("successfull POST request to /api/blogs creates a new blog post and is save
     likes: 12,
   };
 
-  const savedBlog = await api.post("/api/blogs").send(newBlog);
+  const savedBlog = await api.post("/api/blogs").set("Authorization", `Bearer ${auth.token}`).send(newBlog);
 
   const blogsAtEnd = await blogsInDb();
 
@@ -59,7 +73,7 @@ test("if likes is missing from request, it will default to 0", async () => {
     url: "http://www.cs.utexas.edu/~EWD/transcriptions/EWD08xx/EWD808.html",
   };
 
-  const savedBlog = await api.post("/api/blogs").send(newBlog);
+  const savedBlog = await api.post("/api/blogs").set("Authorization", `Bearer ${auth.token}`).send(newBlog);
 
   expect(savedBlog.body.likes).toBe(0);
 });
@@ -72,7 +86,7 @@ describe("if title or url are missing, backend responds with status code 400", (
       likes: 12,
     };
 
-    await api.post("/api/blogs").send(newBlog).expect(400);
+    await api.post("/api/blogs").send(newBlog).set("Authorization", `Bearer ${auth.token}`).expect(400);
   });
 
   test("url missing", async () => {
@@ -81,7 +95,7 @@ describe("if title or url are missing, backend responds with status code 400", (
       author: "Robert C. Martin",
       likes: 10,
     };
-    await api.post("/api/blogs").send(newBlog).expect(400);
+    await api.post("/api/blogs").send(newBlog).set("Authorization", `Bearer ${auth.token}`).expect(400);
   });
 
   test("author missing", async () => {
@@ -90,7 +104,7 @@ describe("if title or url are missing, backend responds with status code 400", (
       url: "http://blog.cleancoder.com/uncle-bob/2017/03/03/TDD-Harms-Architecture.html",
       likes: 10,
     };
-    await api.post("/api/blogs").send(newBlog).expect(400);
+    await api.post("/api/blogs").send(newBlog).set("Authorization", `Bearer ${auth.token}`).expect(400);
   });
 });
 
@@ -98,12 +112,22 @@ describe("deleting blog", () => {
   test("successfull delete will send status code 204", async () => {
     const currentBlogs = await blogsInDb();
     const blogToDelete = currentBlogs[0];
-    await api.delete(`/api/blogs/${blogToDelete.id}`).expect(204);
+    await api.delete(`/api/blogs/${blogToDelete.id}`).set("Authorization", `Bearer ${auth.token}`).expect(204);
   });
 
   test("trying to delete non existing blog will send status code 404", async () => {
     const nonExistingBlogId = await nonExistingId();
-    await api.delete(`/api/blogs/${nonExistingBlogId}`).expect(404);
+    await api.delete(`/api/blogs/${nonExistingBlogId}`).set("Authorization", `Bearer ${auth.token}`).expect(404);
+  });
+
+  test("deleting another user's blog will send status code 401 unauthorized", async () => {
+    const passwordHash = await bcrypt.hash("password", 10);
+    const newUser = new User({ username: "new_user", passwordHash });
+    await newUser.save();
+    const newAuth = await api.post("/api/login").send({ username: "new_user", password: "password" });
+    const currentBlogs = await blogsInDb();
+    const blogToDelete = currentBlogs[0];
+    await api.delete(`/api/blogs/${blogToDelete.id}`).set("Authorization", `Bearer ${newAuth.body.token}`).expect(401);
   });
 });
 
@@ -117,7 +141,7 @@ describe("updating blog", () => {
       url: blogToUpdate.url + "_updated",
       likes: blogToUpdate.likes + 10,
     };
-    await api.put(`/api/blogs/${blogToUpdate.id}`).send(blogWithUpdatedValues).expect(200);
+    await api.put(`/api/blogs/${blogToUpdate.id}`).set("Authorization", `Bearer ${auth.token}`).send(blogWithUpdatedValues).expect(200);
   });
 
   test("title missing will yield status code 400", async () => {
@@ -129,7 +153,7 @@ describe("updating blog", () => {
       url: blogToUpdate.url + "_updated",
       likes: blogToUpdate.likes + 10,
     };
-    await api.put(`/api/blogs/${blogToUpdate.id}`).send(blogWithUpdatedValues).expect(400);
+    await api.put(`/api/blogs/${blogToUpdate.id}`).set("Authorization", `Bearer ${auth.token}`).send(blogWithUpdatedValues).expect(400);
   });
 
   test("url missing will yield status code 400", async () => {
@@ -141,7 +165,7 @@ describe("updating blog", () => {
       url: "",
       likes: blogToUpdate.likes + 10,
     };
-    await api.put(`/api/blogs/${blogToUpdate.id}`).send(blogWithUpdatedValues).expect(400);
+    await api.put(`/api/blogs/${blogToUpdate.id}`).set("Authorization", `Bearer ${auth.token}`).send(blogWithUpdatedValues).expect(400);
   });
 
   test("author missing will yield status code 400", async () => {
@@ -153,7 +177,7 @@ describe("updating blog", () => {
       url: blogToUpdate.url + "_updated",
       likes: blogToUpdate.likes + 10,
     };
-    await api.put(`/api/blogs/${blogToUpdate.id}`).send(blogWithUpdatedValues).expect(400);
+    await api.put(`/api/blogs/${blogToUpdate.id}`).set("Authorization", `Bearer ${auth.token}`).send(blogWithUpdatedValues).expect(400);
   });
 });
